separated FDX and EM4100 decoding

+++ fw/logger.h
@@ -11,7 +11,7 @@
 #include <stdint.h>
 
 void log_startup();
-int log_write_new_entry(uint32_t timestamp, uint32_t uid, uint8_t inout);
+int log_write_new_entry(uint32_t timestamp, uint64_t uid, uint8_t inout);
 
 void log_send_data_via_uart();
